requirements/csv_to_db.txt
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
pymysql>=1.1.0  # For MySQL support